refactor: Move flipper mesh generation to separate class.
