+++ lib/gltf/table-exporter.ts
@@ -33,7 +33,7 @@ import { Table } from '..';
 import { exportGltf } from '../refs.node';
 import { logger } from '../util/logger';
 import { BumperItem } from '../vpt/bumper-item';
-import { FlipperItem } from '../vpt/flipper-item';
+import { Flipper } from '../vpt/flipper/flipper';
 import { IRenderable, RenderInfo } from '../vpt/game-item';
 import { PrimitiveItem } from '../vpt/primitive-item';
 import { RampItem } from '../vpt/ramp-item';
@@ -83,7 +83,7 @@ export class TableExporter {
 			{ name: 'primitives', meshes: Object.values<PrimitiveItem>(this.table.primitives), enabled: !!this.opts.exportPrimitives },
 			{ name: 'rubbers', meshes: Object.values<RubberItem>(this.table.rubbers), enabled: !!this.opts.exportRubbers },
 			{ name: 'surfaces', meshes: Object.values<SurfaceItem>(this.table.surfaces), enabled: !!this.opts.exportSurfaces},
-			{ name: 'flippers', meshes: Object.values<FlipperItem>(this.table.flippers), enabled: !!this.opts.exportFlippers},
+			{ name: 'flippers', meshes: Object.values<Flipper>(this.table.flippers), enabled: !!this.opts.exportFlippers},
 			{ name: 'bumpers', meshes: Object.values<BumperItem>(this.table.bumpers), enabled: !!this.opts.exportBumpers },
 			{ name: 'ramps', meshes: Object.values<RampItem>(this.table.ramps), enabled: !!this.opts.exportRamps },
 			{ name: 'lightBulbs', meshes: this.table.lights.filter(l => l.isBulbLight()), enabled: !!this.opts.exportLightBulbs },
