+++ lib/physics/constants.ts
